Retire the dead code that read the bootinfo sysctl.
The sysctl was removed in c29ba5fe6e3aa26e4fa68b1efa16703b55ac8c05
in 2006.  Recently the dead code started to fail to compile.

NB: very likely many (or most) of the recognized structures defined
in the __DATA__ part of the Makefile.PL are also a dead code.